@@ -1,23 +1,25 @@
 {
   "name": "wasp-ingest-nordic-cloud",
-  "version": "0.0.11",
+  "version": "1.0.0",
   "description": "WASP Ingest using nRF Cloud",
+  "type": "module",
   "main": "app/index.js",
   "scripts": {
     "lint": "eslint .",
     "depcheck": "depcheck",
     "start": "node app/index.js",
     "dev": "nodemon app/index.js | pino-colada",
-    "test": "NODE_ENV=test mocha --config ./test/mocharc.js ./test",
-    "test:integration": "NODE_ENV=test mocha --config ./test/mocharc.js ./test/integration",
-    "coverage": "nyc npm run test"
+    "test": "NODE_ENV=test mocha --config ./test/mocharc.cjs ./test",
+    "test:integration": "NODE_ENV=test mocha --config ./test/mocharc.cjs ./test/integration",
+    "coverage": "c8 npm run test"
   },
   "repository": {
     "type": "git",
     "url": "git+https://github.com/digicatapult/wasp-ingest-nordic-cloud.git"
   },
   "engines": {
-    "node": ">=14.16.0"
+    "node": ">=18.x.x",
+    "npm": ">=9.x.x"
   },
   "keywords": [
     "WASP",
@@ -31,27 +33,26 @@
   },
   "homepage": "https://github.com/digicatapult/wasp-ingest-nordic-cloud#readme",
   "devDependencies": {
-    "chai": "^4.2.0",
-    "depcheck": "^1.3.1",
-    "eslint": "^7.18.0",
-    "eslint-config-prettier": "^7.2.0",
-    "eslint-plugin-prettier": "^3.3.1",
-    "mocha": "^9.1.2",
-    "nodemon": "^2.0.7",
-    "nyc": "^15.1.0",
-    "pino-colada": "^2.1.0",
-    "prettier": "^2.2.1",
-    "supertest": "^6.1.3"
+    "chai": "^4.3.7",
+    "depcheck": "^1.4.5",
+    "eslint": "^8.47.0",
+    "eslint-config-prettier": "^9.0.0",
+    "eslint-plugin-prettier": "^5.0.0",
+    "mocha": "^10.2.0",
+    "nodemon": "^3.0.1",
+    "c8": "^8.0.1",
+    "pino-colada": "^2.2.2",
+    "prettier": "^3.0.2",
+    "supertest": "^6.3.3"
   },
   "dependencies": {
-    "delay": "^5.0.0",
-    "dotenv": "^8.2.0",
-    "envalid": "^7.1.0",
-    "express": "^4.17.1",
-    "kafkajs": "^1.15.0",
-    "mqtt": "^4.2.6",
-    "pino": "^6.11.0",
-    "pino-http": "^5.3.0",
-    "uuid": "^8.3.2"
+    "dotenv": "^16.3.1",
+    "envalid": "^7.3.1",
+    "express": "^4.18.2",
+    "kafkajs": "^2.2.4",
+    "mqtt": "^5.0.3",
+    "pino": "^8.15.0",
+    "pino-http": "^8.4.0",
+    "uuid": "^9.0.0"
   }
 }
