@@ -1,9 +1,9 @@
-const { Kafka, logLevel: kafkaLogLevels } = require('kafkajs')
-require('dotenv').config()
-const mqtt = require('mqtt')
-const delay = require('delay')
+import { Kafka, logLevel as kafkaLogLevels } from 'kafkajs'
+import 'dotenv/config'
+import mqtt from 'mqtt'
+import { setTimeout } from 'node:timers/promises'
 
-const env = require('../../../app/env')
+import env from '../../../app/env.js'
 
 const createPubSub = async () => {
   const mqttClient = mqtt.connect(env.NORDIC_CLOUD_MQTT_ENDPOINT)
@@ -33,15 +33,15 @@ const createPubSub = async () => {
     const prefix = env.NORDIC_CLOUD_MQTT_TOPIC_PREFIX
     mqttClient.publish(`${prefix}/nrf-111111111111111/d2c`, message)
     while (messages.length < expectedMessages) {
-      await delay(10)
+      await setTimeout(10)
     }
-    await delay(waitForExcessMessagesMS)
+    await setTimeout(waitForExcessMessagesMS)
     return [...messages]
   }
 
   return new Promise((resolve) => {
     while (!mqttClient.connected) {
-      delay(50)
+      setTimeout(50)
     }
 
     resolve({
@@ -54,4 +54,4 @@ const createPubSub = async () => {
   })
 }
 
-module.exports = createPubSub
+export default createPubSub
